chore: move browser data utils to other file
