+++ src/routes/[short_id]/+page.svelte
@@ -1,56 +1,15 @@
 <script lang="ts">
 	import { onMount } from 'svelte';
-	import type { PageData } from './$types';
 	import { type RedirectCreate, createRedirect } from '$lib/api/redirects';
-
-	export let data: PageData;
-
-	function getBrowserInfo(userAgent): string {
-		const browserMap = [
-			{ name: 'Firefox', keywords: ['Firefox'] },
-			{ name: 'Opera', keywords: ['Opera', 'OPR'] },
-			{ name: 'Internet Explorer', keywords: ['Trident'] },
-			{ name: 'Edge', keywords: ['Edge'] },
-			{ name: 'Chrome', keywords: ['Chrome'] },
-			{ name: 'Safari', keywords: ['Safari'] }
-		];
-
-		for (const { name, keywords } of browserMap) {
-			if (keywords.some((keyword) => userAgent.includes(keyword))) {
-				return name;
-			}
-		}
-
-		return 'Unknown';
+	import { sendRedirectData } from '$lib/helpers/browserData';
+	export let data: RedirectPageData;
+	interface RedirectPageData {
+		link_id: string;
+		redirect_url: string;
+		detail: string | null;
 	}
 
-	async function getIpAddress(): string {
-		try {
-			const response = await fetch('https://api.ipify.org?format=json');
-			const data = await response.json();
-			return data.ip;
-		} catch (error) {
-			console.log('Error fetching IP address:', error);
-			return null;
-		}
-	}
-
-	async function sendRedirectData() {
-		const current_ip = await getIpAddress();
-		const browser_name = getBrowserInfo(navigator.userAgent);
-
-		const newRedirect: RedirectCreate = {
-			link_id: data.link_id,
-			ip: current_ip,
-			user_agent: navigator.userAgent,
-			referrer: document.referrer,
-			browser: browser_name,
-			platform: navigator.platform,
-			language: navigator.language
-		};
-
-		await createRedirect(newRedirect);
-	}
+	// sendRedirectData();
 
 	onMount(() => {
 		if (data.detail === null) {
